+++ cli/apkInspector.py
@@ -6,8 +6,18 @@ from apkInspector import __version__ as version
 from apkInspector.extract import extract_file_based_on_header_info, extract_all_files_from_central_directory
 from apkInspector.headers import find_eocd, parse_central_directory, headers_of_filename, print_headers_of_filename, \
     get_and_save_local_headers_of_all, show_and_save_info_of_central
-from apkInspector.helpers import save_data_to_file
+from apkInspector.helpers import save_data_to_file, pretty_print_header
+from apkInspector.indicators import apk_tampering_check
 from apkInspector.manifestDecoder import get_manifest
+
+
+def print_nested_dict(dictionary, parent_key=''):
+    for key, value in dictionary.items():
+        if isinstance(value, dict):
+            print_nested_dict(value, parent_key=f"{parent_key}")
+        else:
+            full_key = f"{parent_key}->{key}" if parent_key else key
+            print(f"{full_key}: {value}")
 
 
 def main():
@@ -31,6 +41,8 @@ def main():
     parser.add_argument('-m', '--manifest', action='store_true',
                         help='Extract and decode the AndroidManifest.xml')
     parser.add_argument('-sm', '--specify-manifest', help='Pass an encoded AndroidManifest.xml file to be decoded')
+    parser.add_argument('-a', '--analyze', action='store_true',
+                        help='Check an APK for static analysis evasion techniques')
     parser.add_argument('-v', '--version', action='store_true', help='Retrieves version information')
     args = parser.parse_args()
 
@@ -54,7 +66,8 @@ def main():
     if args.apk is None and args.specify_manifest is None:
         parser.error('APK file or AndroidManifest.xml file is required')
     if not (args.specify_manifest is None) != (args.apk is None):
-        parser.error('Please specify an apk file with flag "-apk" or an AndroidManifest.xml file with flag "-sm", but not both.')
+        parser.error(
+            'Please specify an apk file with flag "-apk" or an AndroidManifest.xml file with flag "-sm", but not both.')
     if args.apk:
         apk_name = os.path.splitext(args.apk)[0]
         with open(args.apk, 'rb') as apk_file:
@@ -66,17 +79,19 @@ def main():
 
             if args.filename and args.extract:
                 try:
-                    cd_h_of_file, local_header_of_file = headers_of_filename(apk_file, central_directory_entries, args.filename)
+                    cd_h_of_file, local_header_of_file = headers_of_filename(apk_file, central_directory_entries,
+                                                                             args.filename)
                 except TypeError as e:
                     print(f"Are you sure the filename: {args.filename} exists?")
                     exit()
                 offset = cd_h_of_file["Relative offset of local file header"]
                 print_headers_of_filename(cd_h_of_file, local_header_of_file)
-                extracted_data = extract_file_based_on_header_info(apk_file, offset, local_header_of_file)
+                extracted_data = extract_file_based_on_header_info(apk_file, offset, local_header_of_file)[0]
                 save_data_to_file(f"EXTRACTED_{args.filename}", extracted_data)
             elif args.filename:
                 try:
-                    cd_h_of_file, local_header_of_file = headers_of_filename(apk_file, central_directory_entries, args.filename)
+                    cd_h_of_file, local_header_of_file = headers_of_filename(apk_file, central_directory_entries,
+                                                                             args.filename)
                 except TypeError as e:
                     print(f"Are you sure the filename: {args.filename} exists?")
                     exit()
@@ -86,24 +101,41 @@ def main():
                 if not extract_all_files_from_central_directory(apk_file, central_directory_entries, apk_name):
                     print(f"Extraction successful for: {apk_name}")
             elif args.list_local:
-                get_and_save_local_headers_of_all(apk_file, central_directory_entries, apk_name, args.export)
+                get_and_save_local_headers_of_all(apk_file, central_directory_entries, apk_name, args.export, True)
                 print(f"Local headers list complete. Export: {args.export}")
             elif args.list_central:
-                show_and_save_info_of_central(central_directory_entries, apk_name, args.export)
+                show_and_save_info_of_central(central_directory_entries, apk_name, args.export, True)
                 print(f"Central header list complete. Export: {args.export}")
             elif args.list_all:
-                show_and_save_info_of_central(central_directory_entries, apk_name, args.export)
-                get_and_save_local_headers_of_all(apk_file, central_directory_entries, apk_name, args.export)
+                show_and_save_info_of_central(central_directory_entries, apk_name, args.export, True)
+                get_and_save_local_headers_of_all(apk_file, central_directory_entries, apk_name, args.export, True)
                 print(f"Central and local headers list complete. Export: {args.export}")
             elif args.manifest:
                 cd_h_of_file, local_header_of_file = headers_of_filename(apk_file, central_directory_entries,
                                                                          "AndroidManifest.xml")
                 offset = cd_h_of_file["Relative offset of local file header"]
-                extracted_data = io.BytesIO(extract_file_based_on_header_info(apk_file, offset, local_header_of_file))
+                extracted_data = io.BytesIO(
+                    extract_file_based_on_header_info(apk_file, offset, local_header_of_file)[0])
                 manifest = get_manifest(extracted_data)
                 with open("decoded_AndroidManifest.xml", "w", encoding="utf-8") as xml_file:
                     xml_file.write(manifest)
                 print("AndroidManifest was saved as: decoded_AndroidManifest.xml")
+            elif args.analyze:
+                tamperings = apk_tampering_check(apk_file)
+                if tamperings['zip tampering']:
+                    print(
+                        f"\n{len(tamperings['zip tampering'])} file(s) listed below appear to have a tampered zip structure!\n")
+                    for val in tamperings['zip tampering']:
+                        pretty_print_header(val)
+                        for dtls in tamperings['zip tampering'][val]:
+                            print(f"{dtls:40}: {tamperings['zip tampering'][val][dtls]}")
+                else:
+                    print(f"No files were detected were a tampering in the zip structure was present.")
+                if tamperings['manifest tampering']:
+                    print(f"\n\nThe AndroidManifest.xml file was tampered using the following patterns:\n")
+                    print_nested_dict(tamperings['manifest tampering'])
+                else:
+                    print(f"The AndroidManifest.xml file does not seem to be tampered structurally.")
             else:
                 parser.print_help()
     elif args.specify_manifest: